Rename `o.g.a.i.lambdas.{Lambdas => SerializableLambdas}`
